Move the identifier grammar matrix to the shared test home
The accepted/rejected identifier parametrizations exercise the raw
validator, which lives in mcp.shared.extension; the server suite keeps
its class-level validation tests.
