+++ tests/server/mcpserver/test_extension.py
@@ -27,7 +27,6 @@ from mcp.server.extension import (
     ResourceBinding,
     ToolBinding,
     compose_tool_call_interceptor,
-    validate_extension_identifier,
 )
 from mcp.server.mcpserver import Context, MCPServer, require_client_extension
 from mcp.server.mcpserver.resources import TextResource
@@ -361,50 +360,6 @@ async def test_version_pinned_method_is_method_not_found_at_a_disallowed_version
 
     assert exc_info.value.code == METHOD_NOT_FOUND
     assert exc_info.value.error.data == "com.example/pinned"
-
-
-@pytest.mark.parametrize(
-    "identifier",
-    [
-        "io.modelcontextprotocol/ui",
-        "com.example/my_ext",
-        "com.x-y.z2/n.a-b_c",
-        "example/x",
-        "a/b",
-        "com.example/9start",
-    ],
-)
-def test_grammar_conformant_extension_identifiers_are_accepted(identifier: str) -> None:
-    """Spec `_meta` key grammar: dot-separated labels (letter start, letter/digit end,
-    hyphens interior), a slash, then a name that starts and ends alphanumeric."""
-    validate_extension_identifier(identifier, owner="T")
-
-
-@pytest.mark.parametrize(
-    "identifier",
-    [
-        "noprefix",
-        "-foo/bar",
-        ".leading/x",
-        "a..b/x",
-        "foo-/x",
-        "9foo/x",
-        "foo/-bar",
-        "foo/bar-",
-        "foo/",
-        "/bar",
-        "foo/ba r",
-        "io.modelcontextprotocol/ui\n",
-        "",
-        None,
-        42,
-    ],
-)
-def test_malformed_extension_identifiers_are_rejected(identifier: Any) -> None:
-    """Spec `_meta` key grammar: malformed prefixes (bad label start/end, empty labels)
-    and malformed names are rejected, as are non-strings."""
-    with pytest.raises(TypeError):
-        validate_extension_identifier(identifier, owner="T")
 
 
 @pytest.mark.parametrize("method", ["tools/list", "completion/complete"])
